pyproject.toml
[project]
name = "portfolio-ml-system"
version = "1.0.0"
description = "Production ML system with credit risk, fraud detection, price prediction, and demand forecasting"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    # ML Frameworks
    "torch>=2.2",
    "xgboost>=2.0",
    "lightgbm>=4.0",
    "scikit-learn>=1.4",

    # Data
    "pandas>=2.1",
    "numpy>=1.26",
    "faker>=22.0",

    # Experiment Tracking
    "wandb>=0.16",

    # Config
    "pyyaml>=6.0",

    # Visualization
    "matplotlib>=3.8",
    "plotly>=5.18",

    # Web UI
    "gradio>=4.15",

    # API
    "fastapi>=0.109",
    "uvicorn>=0.27",

    # Utilities
    "rich>=13.7",
    "tqdm>=4.66",
    "joblib>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
